pyproject.toml
[tool.poetry]
name = "geoip_api"
version = "1.0.0"
description = "This API adds, deletes and provides the geolocation of a given ip or url."
authors = ["Goncalo Moura"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
uvicorn = {extras = ["standard"], version = "^0.18.1"}
pydantic = "^1.9.1"
SQLAlchemy = "^1.4.38"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.5"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
PyMySQL = "^1.0.2"
robotframework = "^5.0.1"
robotframework-seleniumlibrary = "^6.0.0"
robotframework-requests = "^0.9.3"
robotframework-jsonlibrary = "^0.4.1"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
datamodel-code-generator = "^0.13.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
